3 course/5 semester/Web-Tech-2/lab_5/spring-core-task-task/src/main/java/com/epam/rd/autotasks/Task.java
package com.epam.rd.autotasks;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Component;

@Component
public class Task {

    private String description;

    @Autowired
    @Qualifier("assignee")
    private Employee assignee;

    @Autowired
    @Qualifier("reviewer")
    private Employee reviewer;

    public Task() { }

    @Autowired
    public Task(String description) {
        this.description = description;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Employee getAssignee() {
        return assignee;
    }

    public void setAssignee(Employee assignee) {
        this.assignee = assignee;
    }

    public Employee getReviewer() {
        return reviewer;
    }

    public void setReviewer(Employee reviewer) {
        this.reviewer = reviewer;
    }
}
